refactor(store): improve selectors & actions
