+++ src/store/conversionSettings/conversionSettings.actions.js
@@ -1,13 +1,6 @@
 export const SET_CONVERSION_SETTING = 'conversionSettings/SET_CONVERSION_SETTING';
 export const RESET_CONVERSION_SETTINGS = 'conversionSettings/RESET_CONVERSION_SETTINGS';
 
-export const CONVERSION_SETTINGS = {
-    audioBitrate: 'audioBitrate',
-    audioChannels: 'audioChannels',
-    audioCodec: 'audioCodec',
-    audioSampleRate: 'audioSampleRate',
-};
-
 export const resetConversionSettings = dispatch => dispatch({ type: RESET_CONVERSION_SETTINGS });
 export const setConversionSetting = (settingName, settingValue) => dispatch =>
     dispatch({ settingName, settingValue, type: SET_CONVERSION_SETTING });
