+++ src/store/conversionSettings/conversionSettings.constants.js
@@ -1,0 +1,8 @@
+/* eslint-disable import/prefer-default-export */
+
+export const CONVERSION_SETTINGS = {
+    audioBitrate: 'audioBitrate',
+    audioChannels: 'audioChannels',
+    audioCodec: 'audioCodec',
+    audioSampleRate: 'audioSampleRate',
+};
